pyethapp-crawler/requirements.txt
gevent==1.1.0
gipc==0.4.0
CodernityDB
leveldb
lmdb
click
decorator
wheel
pyyaml
werkzeug
ipython>=3.0.0,<5.0.0
statistics
requests
rlp>=0.4.4
devp2p>=0.8.0
ethereum>=1.5.1
pbkdf2
scrypt
pexpect
